Refactor method getRequest and rename it to parseRequest.
The method parseRequest performs blocking/synchrnous IO operations, so
it was refactored accepting the client socket as parameter in order to
not block the main server thread and be run in the client handling
thread.

This name is also more convenient as it better reflect the purpose of
this method which is try to read the request data structure from the
client socket.