Refactor method run in order to solve the server stuck bug.
The server was getting stuck because the method .getRequest() was being
run in the server main thread and as this method performs blocking IO
operations it could stuck the server in case of network latency.

How it was before:

     while (true) try {
      this.getRequest() foreach { req =>
        if (verbose) println(s"\n${new java.util.Date()} - path = ${req.path}  method = ${req.method} - address = ${req.address}")
        Utils.withThread{ this.serveRequest(req)}
     .....

Now it is: The method parseRequest (old getRequest) is run in the
client handler theread.

+
+      val client = this.ssock.accept()
+
+      Utils.withThread{
+        this.parseRequest(client) foreach { req =>
+          if (verbose)
+            println(s"\n${new java.util.Date()} - path = ${req.path} - method = ${req.method} - address = ${req.address}")
+          this.serveRequest(req)
+          println("Client request served")
+        }
       }